Util refactoring, added HTTP submodule, changed url parsing logic
