+++ lib/Proxy.js
@@ -2,6 +2,7 @@
 
 const Http = require('http');
 
+const Boom = require('boom');
 const Co = require('co');
 const Config = require('config');
 const HttpProxy = require('http-proxy');
@@ -13,9 +14,9 @@ const Util = require('./util');
 
 // Internal logic
 
-function handleForbidden(pathname, req, res) {
-  Log.error(`Tried to access forbidden path ${pathname}`);
-  Util.sendServerError(new Error(`Path ${pathname} is forbidden`), req, res);
+function handleForbidden(path, req, res) {
+  Log.error(`Tried to access forbidden path ${path}`);
+  Util.HTTP.sendError(Boom.notFound('Unreachable'), req, res);
 }
 
 class Proxy {
@@ -38,8 +39,8 @@ class Proxy {
           }
         })
         .catch((err) => {
-          Log.error(err, 'Error occurred while handling HTTP request');
-          Util.sendServerError(err, req, res);
+          Log.error(err, `Error occurred while handling HTTP request ${req.method} ${req.url}`);
+          Util.HTTP.sendError(err, req, res);
         });
     });
     this.server.on('upgrade', (req, socket, head) => {
@@ -48,26 +49,19 @@ class Proxy {
     });
     this.forward.on('error', (err, req, res) => {
       Log.error(err, 'Error occurred while proxying request');
-      Util.sendServerError(err, req, res);
+      Util.HTTP.sendError(err, req, res);
     });
   }
 
   handle(req, res) {
     return Co.call(this, function* () {
       const url = Util.parse(req.url);
-      if (Util.match(url.pathname, this.rules.forbid)) {
-        handleForbidden(url.pathname, req, res);
+      if (url.matches(this.rules.forbid)) {
+        handleForbidden(url.path, req, res);
         return;
       }
 
-      if (Util.match(url.pathname, this.rules.ignore)) {
-        if (Util.match(url.pathname, '/**/users/create')) {
-          // Special handling for this rule because we want to allow first time creation but disable all others
-          const load = yield Mattermost.Api.user.initialLoad();
-          if (!load.no_accounts) {
-            handleForbidden(url.pathname, req, res);
-          }
-        }
+      if (url.matches(this.rules.ignore)) {
         return;
       }
