+++ lib/auth/OIDC.js
@@ -39,7 +39,7 @@ class OIDC {
       const url = Util.parse(req.url);
       if (!url.query.code) {
         const authorizationURL = self.getAuthorizationURL(req, res);
-        return Util.redirect(res, authorizationURL);
+        return Util.HTTP.redirect(res, authorizationURL);
       }
 
       const token = yield self.getToken(req, res);
