+++ lib/auth/index.js
@@ -14,5 +14,5 @@ const DESTINATION_COOKIE = Config.get('auth.destination.cookie');
 exports.OIDC = OIDC;
 
 exports.finalize = function(req, res) {
-  Util.redirect(res, Util.Cookie.for(req).get(DESTINATION_COOKIE.name));
+  Util.HTTP.redirect(res, Util.Cookie.for(req).get(DESTINATION_COOKIE.name));
 };
