+++ lib/util/index.js
@@ -2,53 +2,34 @@
 
 const URL = require('url');
 
-const Boom = require('boom');
 const camelCase = require('lodash.camelcase');
 const mapObj = require('map-obj');
 const minimatch = require('minimatch');
 const snakeCase = require('lodash.snakecase');
 
 const Cookie = require('./Cookie');
+const HTTP = require('./HTTP');
 const Iron = require('./Iron');
-const Log = require('../../config/logging').get('UTIL');
 
 // Define exports
 
 exports.Cookie = Cookie;
 
+exports.HTTP = HTTP;
+
 exports.Iron = Iron;
 
-exports.redirect = function(res, url) {
-  res.writeHead(302, {
-    Location: url
-  });
-  res.end();
-};
-
-exports.sendServerError = function(err, req, res) {
-  if (res.finished) {
-    Log.error('Cannot send server error because response is already closed');
-    return;
-  }
-
-  const output = (err.isBoom ? err : Boom.badImplementation('', err)).output;
-  const accept = req.headers.accept;
-  const headers = {};
-  let data;
-  if (accept && accept.toLowerCase().includes('application/json')) {
-    headers['Content-Type'] = 'application/json';
-    data = JSON.stringify(output.payload);
-  } else {
-    headers['Content-Type'] = 'text/plain';
-    data = `${output.payload.statusCode} ${output.payload.message}`;
-  }
-
-  res.writeHead(output.statusCode, output.payload.error, headers);
-  res.end(data);
-};
-
 exports.parse = function(url) {
-  return URL.parse(url, true);
+  const urlObject = URL.parse(url, true);
+  urlObject.matches = function(patterns, path = false) {
+    return exports.match(path ? urlObject.path : urlObject.pathname, patterns);
+  };
+
+  return urlObject;
+};
+
+exports.match = function(path, patterns) {
+  return Boolean(exports.toArray(patterns).find((pattern) => minimatch(path, pattern)));
 };
 
 exports.toArray = function(value) {
@@ -57,10 +38,6 @@ exports.toArray = function(value) {
   }
 
   return [];
-};
-
-exports.match = function(path, patterns) {
-  return Boolean(exports.toArray(patterns).find((pattern) => minimatch(path, pattern)));
 };
 
 exports.toSnakeCase = function(obj) {
